data: Move file review action commands
File-review output renders suggested follow-up commands for page-scoped
selections, and validation state uses the same command text in refusal
messages.

The formatter lives in validation state, so output rendering imports state
validation code for command text that does not need freshness checks.

This commit addresses that by moving file-review action command formatting
into data.file_review_action_commands and by switching state and output
callers to that module.

The next commit will update architecture guards for the new formatter
boundary.
